Major features added -
Implemented Google Firebase Support for Authentication to sign in the user with their Google Account, allowing for their app experience to be personalised with their own trip recommendations shown in the 'Card Carousel' (still in testing)

Implemented Firebase Firestore Support for document querying and data collection, to securely record user login details in a NoSQL database. If the user asks to remove their account details, the owner of the application can remove their details from the Firebase dashboard (still in testing)

Implemented Firebase Cloud Storage to populate the 'card carousel' (in the MyHomeFragment) with images of recommended trips from the storage bucket, shown to users that are logged in (still in development).

Implemented MVVM design pattern to allow for Google Firebase functionality, whilst allowing the UI to respond to live data changes to Firebase using ViewModels and Repositories (still in testing).